refactor(workspace-fs): extract pure event algebra from watch.ts
watch.ts had grown to ~880 lines. The per-path coalescing and
delete/create-to-rename reconciliation are pure functions with no manager
or I/O coupling — move them (and their tests) to watch-event-coalescing.ts,
leaving watch.ts as the FsWatcherManager and its native-subscription
lifecycle.
